+++ BankApp/src/main/java/jkeeler/App/Service/TransactionService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import jkeeler.App.Entity.Transaction;
 
 public interface TransactionService {
-    List<Transaction> findByBankId(int bankId);
+    List<Transaction> findByBankId(Long bankId);
     void addTransaction(Transaction transaction);
     // Add other transaction-related methods as needed
 }
